# controllers/equipment.go
package controllers

import (
	"github.com/gin-gonic/gin"
)

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         equipment
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /equipment [get]
// @Security     ApiKeyAuth
func GetAllEquipment(c *gin.Context) {

}

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         equipment
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /equipment [post]
// @Security     ApiKeyAuth
func CreateEquipment(c *gin.Context) {

}

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         equipment
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /equipment [put]
// @Security     ApiKeyAuth
func UpdateEquipment(c *gin.Context) {

}

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         equipment
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /equipment [delete]
// @Security     ApiKeyAuth
func DeleteEquipment(c *gin.Context) {

}